EC: Move commitments to ec.util
This will enable us to switch between different implementations of
points and scalars
